% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-stratocastR.Rmd.
\name{background_by_treatment}
\alias{background_by_treatment}
\title{Define how treatment conditions should feed into background subtraction, as test wells or as blanks}
\usage{
background_by_treatment(test, blanks, ignore = NULL, treatments)
}
\arguments{
\item{test}{(character) The treatment group that should be corrected by background subtraction;
should match a \code{treatment} value in the plate layout}

\item{blanks}{(character) The treatment group that should be used as the background to subtract;
should match a \code{treatment} value in the plate layout}

\item{ignore}{(character; defaults to NULL) Any treatment group(s) that should be ignored in analysis}

\item{treatments}{(character) Plate layout CSV file, following the \code{make_layout_file()} format}
}
\value{
A dataframe with one row per well and three columns:
\itemize{
\item well:  the well ID, e.g., "A1"
\item bgtype:  the categorical column defining each well as "blank", "test", or "ignore".  Blank wells
are used to determine background; test wells have background subtracted from them; ignore wells
are ignored.
\item bggroup:  a column defining groups within which background subtraction should be performed.
Here, 1 for all test or blank wells, otherwise NA.
}
}
\description{
Define how treatment conditions should feed into background subtraction, as test wells or as blanks
}
\seealso{
background_by_wells() for defining more complex background-subtraction rules with multiple
groups
}