% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-stratocastR.Rmd.
\name{background_by_wells}
\alias{background_by_wells}
\title{Define how wells should feed into background subtraction, as test wells or as blanks}
\usage{
background_by_wells(bgfile)
}
\arguments{
\item{bg}{(character) Background layout file (.csv or .xlsx), following the
\code{make_bg_file()} format. Values in the file should take the form \verb{group;type} (e.g., 1;B,
1;blank, 1;test); the delimiter between group and type can be specified via \code{bgdelim}}
}
\value{
A dataframe with one row per well and three columns:
\itemize{
\item well:  the well ID, e.g., "A1"
\item bgtype:  the categorical column defining each well as "blank", "test", or "ignore".  Blank wells
are used to determine background; test wells have background subtracted from them; ignore wells
are ignored.
\item bggroup:  a column defining groups within which background subtraction should be performed.  E.g.,
if two different media were used in the plate, background subtraction should be performed
separately for wells using each medium.
}
}
\description{
Define how wells should feed into background subtraction, as test wells or as blanks
}
\seealso{
background_by_treatment() for defining simpler background-subtraction rules with a single
group
}