% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-stratocastR.Rmd.
\name{detect_outliers}
\alias{detect_outliers}
\title{Convert a plate dataset to a tsibble; detect outlier timepoints in each well}
\usage{
detect_outliers(
  data,
  interval,
  iter = 1,
  incol = c("absorbance", "abs.corr", "airtemp.C")
)
}
\arguments{
\item{data}{(dataframe) Plate data data.frame, as generated by read_platedata()}

\item{interval}{(numeric) Length of time step in minutes, as read by get_interval_min()}

\item{iter}{(numeric) Number of outlier-detection iterations for \code{forecast::tsoutliers()} to run.
Defaults to 1.}

\item{incol}{(character) Column to analyze for outliers.  To examine raw absorbances, e.g., in
blank wells before aggregation, use "absorbance"; to examine background-subtracted absorbances,
e.g., in test wells, use "abs.corr"; to examine air temperature, use "airtemp.C".}
}
\value{
A tsibble that adds two new columns to data:
\itemize{
\item a per-well per-timestep outlier detection result from \code{forecast::tsoutliers()}; depending on
the value of \code{incol}, the column may be \code{is.outlier.temp} or \code{is.outlier.abs}
\item \code{n.out.bgtype}, a per-timestep count of the number of wells in the background type flagged as
outliers
}
}
\description{
Convert a plate dataset to a tsibble; detect outlier timepoints in each well
}