% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-stratocastR.Rmd.
\name{get_interval_min}
\alias{get_interval_min}
\title{Get the measurement interval length and convert to minutes}
\usage{
get_interval_min(filename)
}
\arguments{
\item{filename}{(character) Cerillo CSV file}
}
\value{
The interval between measurements (in minutes) in the experiment
}
\description{
Get the measurement interval length and convert to minutes
}