% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-stratocastR.Rmd.
\name{make_bg_file}
\alias{make_bg_file}
\title{Create a blank template file for defining background layout}
\usage{
make_bg_file(filename)
}
\arguments{
\item{filename}{(character) Destination filename, with .xlsx or .csv extension}
}
\value{
If filename's extension is .xlsx, an Excel file with two sheets
\itemize{
\item \code{layout}, containing the background layout
\item \code{params}, defining the strings used in the layout
If filename's extension is .csv, a .csv file with header lines defining the strings used
in the layout and a table containing the layout.
}
}
\description{
Create a blank template file for defining background layout
}