% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-stratocastR.Rmd.
\name{make_layout_file}
\alias{make_layout_file}
\title{Create a fresh plate layout file, as .xlsx or .csv.}
\usage{
make_layout_file(filename)
}
\arguments{
\item{filename}{(character) Destination filename, with .xlsx or .csv extension}
}
\description{
Create a fresh plate layout file, as .xlsx or .csv.
}