% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-stratocastR.Rmd.
\name{plot_abs}
\alias{plot_abs}
\title{Plot plate-reader data by well, by row, and by column, including temperature and outlier status.}
\usage{
plot_abs(
  data,
  exclude.outliers = TRUE,
  threshold = 1,
  show.ignores = FALSE,
  dropstyle = c("drop", "pointless", NULL),
  use.corr = TRUE,
  title,
  dates,
  caption = NULL
)
}
\arguments{
\item{data}{(dataframe) Plate-reader data processed through \code{detect_outliers()}}

\item{exclude.outliers}{(logical) Should points identified as outliers be hidden? Defaults to
TRUE.}

\item{threshold}{(integer) What fraction of wells with outliers is too many for the data to be
shown? For timepoints when the fraction of outliers exceeds plot.threshold, all measurements will
be treated as outliers for plotting.  Defaults to 1, i.e., any number of outliers is acceptable.}

\item{show.ignores}{(logical) Should data from wells of bgtype "ignore" be plotted?}

\item{dropstyle}{(character) If outliers are to be dropped, how should they be dropped?  With
\code{dropstyle = "drop"}, outliers are removed from data altogether; lines drawn across the missing
region will be drawn straight between the last point before the gap and the first point after.
With \verb{dropstyle = "pointless", outliers are not plotted as points but are included in the line graph.  }dropstyle\verb{is checked only if}exclude.outliers` is TRUE.}

\item{use.corr}{(logical) For test wells, should background-subtracted (TRUE) or raw (FALSE)
measurements be plotted?  (Raw measurements are always plotted for blanks and, if plotted, ignores.)}

\item{title}{(character) A brief description of the experiment to print as the plot's title.}

\item{dates}{(character) The start date or date range of the experiment, to print as the plot's
subtitle.}

\item{caption}{(character) A caption, e.g., for tracking background-subtraction choices or other
experimental notes.  Defaults to NULL.}
}
\description{
Plot plate-reader data by well, by row, and by column, including temperature and outlier status.
}