% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-stratocastR.Rmd.
\name{process_plate}
\alias{process_plate}
\title{Perform end-to-end plate reader data analysis, optionally saving the resulting plot}
\usage{
process_plate(
  filename,
  treatments,
  background.file = NULL,
  background.treatments = NULL,
  plot.title,
  expt.dates,
  outfile = NULL,
  display.plot = FALSE,
  ts.iter = 1,
  exclude.bg.outliers = TRUE,
  bg.basis = "median",
  bg.threshold = 3,
  show.ignores = FALSE,
  plot.threshold = 1,
  plot.dropstyle = "pointless",
  use.corr = TRUE
)
}
\arguments{
\item{filename}{(character) Cerillo CSV file}

\item{treatments}{(character) Plate layout CSV file, following the \code{make_layout_file()} format}

\item{background.file}{(character) Background layout .csv or .xlsx file, following the
\code{make_bg_file()} format.  Defaults to NULL.  Either \code{background.file} or \code{background.treatments}
must be supplied.}

\item{background.treatments}{(list) Named list of conditions to be handled as test wells, blanks,
and ignores (e.g., \verb{list(test = c("BG.Cab", "Azo"), blank = "Control", ignore = "Empty")).   Defaults to NULL.  Either }background.file\code{or}background.treatments` must be supplied.}

\item{plot.title}{(character) A brief description of the experiment, to print as the title of the
analysis figure.  Other annotations will automatically capture data-analysis parameters;
plot title should briefly describe the experiment.}

\item{expt.dates}{(character) The start date or date range of the experiment, to print as the plot's
subtitle.}

\item{outfile}{(character) Output filename, if analysis figure is to be saved.  Defaults to NULL.}

\item{display.plot}{(logical) Should the analysis figure be displayed on a graphics device?  Can
be slow.  Defaults to FALSE.}

\item{ts.iter}{(numeric) Number of iterations to use in outlier detection.  Defaults to 1.}

\item{exclude.bg.outliers}{(logical) Should outlier readings in blank wells be removed before
aggregating background measurements?  Defaults to TRUE.}

\item{bg.basis}{(character) Should blank wells' absorbance readings be aggregated by mean or median
for background subtraction? Defaults to "median".}

\item{bg.threshold}{(numeric) Minimum number of blank wells with data that are required for
background subtraction to proceed in a background group at a given timepoint?  Defaults to 3.}

\item{show.ignores}{(logical) Should data from wells of the "ignore" type be plotted?  Defaults to
FALSE.}

\item{plot.threshold}{(numeric) What fraction of wells with outliers is too many for the data to be
shown? For timepoints when the fraction of outliers exceeds plot.threshold, all measurements will
be treated as outliers for plotting.  Defaults to 1, i.e., any number of outliers is acceptable.}

\item{plot.dropstyle}{(character) How should outliers be plotted?  They can be dropped altogether,
from points and lines ("drop"); they can be dropped from the points but kept in the lines,
so that they are visible but deemphasized ("pointless"); or they can be included fully (NULL).}

\item{use.corr}{(logical) Should background-subtracted or raw values be shown in test wells?  When
\code{use.corr = TRUE}, background-subtracted values are shown for test wells, while raw values are
shown for blank wells (and "ignore" wells, if plotted), with the aggregate background measure
overlaid on these wells for a visual reference.  When \code{use.corr = FALSE}, only raw values are
shown across the entire plate.}
}
\value{
A list of length 2, containing
\itemize{
\item a dataframe of processed data
\item a patchwork plot object
Additionally, if \code{display.plot} is TRUE, the analysis figure is displayed on a graphics device;
if an output filename is supplied, the analysis figure is saved to disk.
}
}
\description{
Perform end-to-end plate reader data analysis, optionally saving the resulting plot
}