% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-stratocastR.Rmd.
\name{read_platedata}
\alias{read_platedata}
\title{Read in the data file and optionally merge in sample info.}
\usage{
read_platedata(filename, treatments = NULL)
}
\arguments{
\item{filename}{(character) Cerillo CSV file}

\item{treatments}{(character) Plate layout CSV file, following the \code{make_layout_file()} format}
}
\value{
A tidy dataframe with one row for each well's data and metadata at each timepoint
}
\description{
Read in the data file and optionally merge in sample info.
}