% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-stratocastR.Rmd.
\name{read_treatments}
\alias{read_treatments}
\title{Read in the treatment file}
\usage{
read_treatments(treatments)
}
\arguments{
\item{filename}{(character) Plate layout file (.csv, .xlsx), following the
\code{make_layout_file()} format}
}
\value{
A tidy dataframe with one row for each well and two columns:
\itemize{
\item well:  the well ID, e.g., "A1"
\item treatment:  the experimental condition in the well
}
}
\description{
Read in the treatment file
}