% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-stratocastR.Rmd.
\name{subtract_background}
\alias{subtract_background}
\title{Perform background subtraction on a plate reader data file.}
\usage{
subtract_background(
  data,
  bg,
  iter = 1,
  basis = c("mean", "median"),
  interval = NULL,
  exclude.outliers = TRUE,
  threshold = 3
)
}
\arguments{
\item{data}{(dataframe) The output of \code{read_platedata()}; may additionally have been processed through
\code{detect_outliers()}}

\item{bg}{(dataframe) The output of \code{background_by_treatment()} or \code{background_by_well()}}

\item{iter}{(numeric) Number of outlier-detection iterations for \code{forecast::tsoutliers()} to run.
Defaults to 1.}

\item{basis}{(character) Should background values be aggregated by mean or by median before
subtraction?}

\item{interval}{(numeric) Length of time step in minutes, as read by \code{get_interval_min()}; required
if \code{exclude.outliers} is TRUE.}

\item{exclude.outliers}{(logical) Should background values flagged as outliers be excluded from
calculation of background?  Applies only if \code{data} has been processed through \code{detect_outliers()}.
Defaults to TRUE.}

\item{threshold}{(numeric) The minimum number of blank wells allowed to be used for background
subtraction at each timepoint.  Defaults to 3.}
}
\value{
A dataframe that adds three new columns to data:
\itemize{
\item bgtype:  the categorical column defining each well as "blank", "test", or "ignore".  Blank wells
are used to determine background; test wells have background subtracted from them; ignore wells
are ignored.  Merged in from bg.
\item bggroup:  a column defining groups within which background subtraction should be performed.  E.g.,
if two different media were used in the plate, background subtraction should be performed
separately for wells using each medium.  Merged in from bg.
\item bg.n:  a count of blank values aggregated into each group's background value at each time.
\item abs.corr:  a column of background-subtracted values, absorbance - background.  For "blank" and
"test" wells, these values will be the per-well per-time absorbance minus the group's
\code{basis}-aggregated blank absorbance; for "ignore" wells, NA.
\item flag:  notes on the background processing applied to obtain each row's abs.corr value
}
}
\description{
Perform background subtraction on a plate reader data file.
}